requirements.txt
opencv-python>=4.8.0,<4.10
numpy>=1.24.0,<2.0
mss>=9.0.0
easyocr>=1.7.0
Pillow>=10.0.0
matplotlib>=3.7.0
